+++ ProjectA/ProjectA/Models/PlayersModels/ElementsData.cs
@@ -1,8 +1,4 @@
-﻿using Newtonsoft.Json;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+﻿using System.Collections.Generic;
 
 namespace ProjectA.Models.PlayersModels
 {
