+++ ProjectA/ProjectA/Program.cs
@@ -1,11 +1,8 @@
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
-using Microsoft.Extensions.Logging;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+using ProjectA.HostedService;
 
 namespace ProjectA
 {
@@ -21,6 +18,14 @@ namespace ProjectA
                 .ConfigureWebHostDefaults(webBuilder =>
                 {
                     webBuilder.UseStartup<Startup>();
-                });
+                })          
+            .ConfigureServices(services =>
+            {
+                var Configuration = services.BuildServiceProvider().GetRequiredService<IConfiguration>();
+                
+                services.AddHostedService<TelegramBotHostedService>();
+                
+            });
+          
     }
 }
