+++ ProjectA/ProjectA/Services/Handlers/HandlerTeamService.cs
@@ -1,7 +1,4 @@
-﻿using ProjectA.Models.Teams;
-using ProjectA.Services.Teams;
-using System;
-using System.Collections.Generic;
+﻿using ProjectA.Services.Teams;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
