+++ ProjectA/ProjectA/Services/Handlers/IHandlerTeamService.cs
@@ -1,7 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+﻿using System.Threading.Tasks;
 
 namespace ProjectA.Services.Handlers
 {
