+++ ProjectA/ProjectA/appsettings.json
@@ -1,6 +1,6 @@
 {
   "FantasyPremierLeagueUrl": "https://fantasy.premierleague.com",
-  "TelegramBotToken": "1993306232:AAE_tM2iqnfW6njaO_L0At5i7qvGlRcGwMI",
+  "TelegramBotToken": "Your bot token here",
   "Logging": {
     "LogLevel": {
       "Default": "Information",
